Gap buffer of tokens
Now tk.bas uses a gap buffer instead of a simple array. At the same time
it now keeps track of the current position, so there are relative
move/in/out operations. Access to tokens is still globally indexed.

This should help a lot with performance with lots of tokens. This is needed
for huge headers, otherwise every single insert/delete would move KiBs of
memory...