Experimental EXTERN block parsing
Detecting the block, and even the corresponding closing '}' is possible,
without thinking about #ifs for now, and it can be translated, but the
new END EXTERN tokens get in the way of the rest of toplevel parsing.

Either it will be necessary to mark tokens as "already handled", or parse
and translate separately.