Re-add two-pass parsing for better handling of unknown constructs
- 1st pass marks tokens in unknown constructs as "poisoned"
- 2nd pass looks out for poisoned tokens, and turns them into TODOs,
  and merges anything else into the proper high-level tokens

This is again similar to the "marking" done by the old fbfrog parser,
except that only a "poisoned" boolean is used now to mark bad tokens,
and more complex marks aren't needed since both passes now use the same
code, and there is no duplication anymore.

This allows the 2nd parsing pass to insert/delete tokens as it's parsing,
because it can assume that the construct will turn out to be correct,
otherwise the 1st pass would have marked it as poisoned. This is especially
nice for parsing function declarations, because the parameters can be
translated one by one with no worry about having to revert everything if
there is a parsing error during the last parameter or similar.