pp: Don't accept identifiers as atoms in #if expressions
In the CPP any unknown identifiers are normally evaluated as 0, but this
feature is probably rarely used in C headers for which FB bindings should
be created. If ppExpression() encounters this situation the real reason
probably is that the identifier is a #define that wasn't expanded yet.
In that case it's better to fail parsing and treat the #if as an unknown
PP directive. The identifier should be expanded before the #if expression
parsing, if needed it can even be expanded to 0.