pp: Eval/expand unconditionally only, no more trail code duplication
Instead of turning this:
	#ifdef foo
		#define f f1
	#else
		#define f f2
	#endif
	void f(void);
into this:
	#ifdef foo
		#define f f1
		void f(void);
	#else
		#define f f2
		void f(void);
	#endif
to handle expansion of conditionally defined macros, it seems better to
simply expect any #ifs to be solved out by presets to remove any ambiguity,
and parse the same header multiple times, each with different pre-#defines
as needed to extract the wanted declarations.

The problem with the trail-code-duplication is that it has to be done for
every single #if containing a precious #define or #undef, before #if
evaluation can be done, and that seriously blows up the tk buffer size
even for medium sized headers if they use lots of #ifs. Most of the
duplicate code will be solved out by #if evaluation but nevertheless it
slows down the preprocessing step a lot.

Normally a binding only really needs to be parsed on "Linux" or "Win32"
modes, i.e. a lot of uninteresting cases can be ignored immediately.

Of course fbfrog should still support leaving in unsolvable #if blocks,
whose conditions couldn't be evaluated. If a precious #define or #undef is
found in an unsolvable #if block, fbfrog needs to oops() because the preset
didn't do its job of specifying the proper symbols needed to solve all
important #if blocks, and now fbfrog can't know whether that precious
#define or #undef will be reached or not. Checking for this is what makes
it safe to do the macro expansion this way.

Multiple completely separate parsing passes is exactly what will be done
to handle multiple library versions, and the same kind of AST merging is
needed afterwards, so unlike the trail-code-duplication, no extra merging
code is needed.

- Duplicate #defines will now be checked for equality, oops() in case the
  2nd #define is incompatible to the 1st (gcc would warn about
  incompatible redefinition)