cpp: Change strategy, assume unknown symbols = undefined
It's too much work to add a DEFINE or UNDEF for all pre-defines
that a C header checks for (e.g. libpng has hundreds, and changes
with each version).

Instead, fbfrog should immitate gcc, have the same pre-defines,
and allow presets to add additional pre-defines if needed.
-v verbose output reports unsafe assumptions (i.e. if an unknown
symbol is treated as undefined).