ast: Fix version block matching
For example, for astAddVersionedChild():
	existing:
		version "1" or "2"
			foo
	new:
		version "1"
			bar
Technically "1" matches "1" or "2", but that's not the kind of
matching desired here, because adding bar to the same version as
foo here would be wrong, as bar is not available under "1" or "2",
but just "1".

Because of that, a more exact matching function must be used;
for now astIsEqualDecl() (same as used before version matching
overhaul) should be good enough even though it would treat
<"1" or "2"> and <"2" or "1"> as being a mismatch.

Note that not just the OR is affected, also the '.' member op:
	existing:
		version "1".win32
			foo
	new:
		version "1"
			bar
or:
	existing:
		version "1"
			foo
	new:
		version "1".win32
			bar
in both cases, bar can't just be inserted at the same version as
foo.

The same principle is needed for astRemoveVerBlockWrapping()
aswell, only astGet1VersionOnly() can do the more relaxed/soft
matching.