ast: Add extra function for CPP expression folding & improve semantics
Now, it does astFoldKnownDefineds() -> astFold() -> astFoldUnknownIds() ->
astFold(). This way it will ensure to fold as much as it can before solving
out unknown ids and showing the corresponding "assuming undefined"
warnings.

Previously, it would do astFoldUnknownIds() after the 1st pass in the
astFold() loop - which may not have solved out as much as it could,
potentially causing unnecessary "assuming defined" warnings to be shown.