Remove support for multiple root files & -common
i.e. root files won't be treated as separate APIs anymore.
The version-specific parsing process now produces one AST per version and
that's it. Easier to merge aswell. This also makes -common obsolete.
This matches the idea that fbfrog works like a compiler that creates one
binding for exactly the input it's given.