Add -reservedid option to allow triggering symbol renames manually
This allows symbols to be renamed manually, which is sometimes needed to
fix inter-header conflicts.

For example, the Windows API headers have some very generic #defines which
can easily cause problems when mixed with other code or bindings in FB,
due to the case-insensitivity.

GDI #defines:
	#define TRANSPARENT 1
	#define OPAQUE 2

Other code:
	type Image
		transparent as integer  '' unwanted error
	end type

There it makes sense to use "-reservedid transparent", either when creating
one binding or when creating the other, in order to force either the
#define or the field to be renamed.