tk: Increase buffer size exponentially to reduce amount of total resizes
This greatly reduces the amount of memmove()'ing done for tk gap buffer
resizing with huge input such as MinGW-w64's windows.h with all its
dependencies.

In that test case (MinGW-w64's windows.h), this reduces the total runtime
by 50% on my laptop:
	previously: ~20 seconds
	now: ~9 seconds

That's with 4 CPP runs, i.e. 32bit/64bit and UNICODE/non-UNICODE,
and also -filterout '*', which means that almost all the time is spent
in the CPP.