Re-add -include option for specifying pre-#includes
This is useful for inserting additional code which may be needed to parse
the main input files. Pre-#includes themselves do not become part of the
output though.

For example, the libpng headers require some #defines from limits.h.
For this purpose fbfrog provides its own small limits.h (which works for
all of the default targets), and now it becomes possible to use it by
doing "-include limits.h". This allows the libpng headers to be parsed.