Preserve unexpanded direct #includes and move them to top
This changes the handling of #include statements. Previously they were
expanded if the file was found, and not expanded if the file was not found.
Now they will be preserved depending on whether the include file content
will be preserved or not.

I.e. #includes for files which fbfrog can find but which will be filtered
out due to -filterout won't be missing anymore.