Shorten TODO notes
The TODO notes emitted into the output .bi are now much shorter - just
single lines. The fbfrog parser error message isn't shown here anymore;
instead it's printed to console under -v.

This makes the TODO notes much easier to read and understand. The parser
error message is usually only useful for debugging fbfrog itself, but not
when working on a binding. All that matters is whether fbfrog could do
the auto-translation or not.