Forget about dropping include guard #defines
The #defines left behind by include guards are generally "ugly", and
because of that it would prefer to exclude them from .bi files, but on the
other hand there are cases where C headers check for others having been
included by checking for that include guard #define.

For example:
	#ifdef FOO_H
		// Expose foo-specific parts of the API
	#endif

In order to support making bindings for such cases, the include guard
#defines should be preserved... of course this could be handled differently
too, especially when .bi files are adjusted manually, but generally
speaking, this shows why it's dangerous to remove/rename symbols from
headers.