Revert "Handle C++ bool type just like C _Bool"
This reverts commit 50a51d715f0badf0430d7d9836ef0c0237db62b2. In C mode,
"bool" isn't a built-in type but just a #define from stdbool.h, and in fact
some C libraries define their own "bool" type. Because of that we can't
treat it as the built-in type in C code. If a C++ mode was ever added to
fbfrog, "bool" should be recognized there though.