c: Delay parsing of #define bodies until EOF
This is an alternative to the typedef preparsing which was removed in
commit 9c8af6957ce28d0a2a637ba74b275143d3b735e2.

#define body aren't parsed immediately, but just collected in a list.
After the C parser has parsed all the input (and seen all declarations)
it will process the list and parse #define bodies. By then it will have
seen all typedefs in the input.

As an optimization, this is only done for #define bodies which contain
identifiers, but not for example for simple number constants (which are
very common, while #defines containing identifiers are relatively rare).