Turn -rename*/-remove*/-typedefhint/etc. into global options
It doesn't really seem very useful to make them version-specific.
If we're renaming or removing a symbol, it's like going to be needed in
all versions. Or maybe it's only needed for one version, and is a no-op
for the others. Either way, this simplifies the code a bit and makes
fbfrog.bas:hApplyOptions() unnecessary, which was rather ugly (and slow).