c: Improve handling of declarators with function type
Most importantly, this improves the declarator parsing to handle
anonymous parameters with function type. For example:
	void f(  int ()    );
f is a procedure which takes a parameter with function type. The
parameter is a function returning an int and "empty" parameter list,
though it really becomes a function pointer. (similar to how array
parameters decay to pointers and are implicitly passed byref)
Previously it was handled as simple anonymous parameter of type int,
with the parentheses enclosing the non-existing identifier.

More examples of parameters with function type:
	void f(  int (())       );
	void f(  int (int)      );
	void f(  int ((int))    );

Furthermore, function typedefs are now allowed to be expanded into
parameters, and all parameters with function type are automatically
turned into function pointers.