Remove support for verification code for -declareversions/-declaredefines
Generally, if such verification code is needed in a binding, it can be
added manually - the automation doesn't help much here. In fact when
translating multiple files that are supposed to be used together, then it
doesn't even make sense to have the verification code in every single one.
Then it would have to be removed manually from every file - at that point
it's just better not to auto-generate it at all.