error reporting: Track token's "real" location, forget about offset/length
This changes the CPP to assign the macro call location to the
corresponding expansion tokens. This way we can report those as appearing
at the macro call site, which is in many cases more useful than reporting
the error in the #define body. Also, fbfrog will now only report the
file name and line number, and won't show the line of source code anymore.
(but the tokens will be shown instead)

Overall, the error reports should be shorter while staying as useful, and
the code to implement it should be simpler.