Return to "old" modular design, re-add highlevel.bas
I think it's best to return fbfrog to its "old" modular design, where
C parsing and the various "highlevel" passes for the conversion to FB
were separate, and could easily be added/removed. The current design tries
to do as much as possible in one go (during the C parsing), which may be a
bit faster, but it's also much more complex and less flexible.

This basically reverts 84287e4e70 and some others.
* Almost all the highlevel operations are turned back into separate passes
* ASTNODEs are self-contained again, astDelete() is back and improved,
  no more cross-referencing, no more memory "leaks". astInsert() improved.
* C parser functions return ASTNODEs again instead of building the AST via
  a stack
* Added astVisit() to simplify many of the highlevel passes