Add -emit, remove -filter*
The idea of fbfrog only emitting one file at a time, and the user using
-filterout/-filterin options to control the content of that file, doesn't
scale well at all if we have big enough header set and want to translate
most files 1:1 instead of merging all into 1 (i.e. Windows API headers).

It means that fbfrog has to parse the whole "set of header files" again
and again, for every .bi file that should be produced. It's bad enough that
we have to parse multiple times already due to multiple targets/versions,
but essentially it becomes a bad case of O(n^2).

Thus, this changes the tool to a different idea: load the API "once",
and specify what files to emit (and if needed, what goes into each file).

* -filterout/-filterin removed, -emit added
* #includes are now preserved in-place if needed
* forward declarations are now added in-place
* highlevel work split into two procedures, one for the "global" AST work,
  one for the .bi-file-specific AST work
* source locations are preserved on ASTNODEs again, then later the
  declarations are distributed into .bi files based on that.