Add -dontemit option
This allows adding some patterns for "not emitting" in front of more
generic -emit options. For example:
	-dontemit '*/foo.h' -emit '*' all.bi
The -emit will match everything, except the files that matched some
previous pattern (no matter whether that was an -emit or a -dontemit).

This is basically reimplementing the old -filterout idea with the new -emit
system.