internal: Split SOURCEBUFFER => SourceInfo/FileBuffer
This makes lexLoad* to parse input buffers directly, and treats the source
information separately. The input buffer is no longer preserved forever as
far as TkLocation/SourceInfo are concerned - of course FileBuffers are
still cached, but in the other cases where we lexLoadC() from temporary
strings, those temporary strings won't be duplicated anymore.