Make all binding generation options API-specific
This allows passing -removedefine etc. conditionally, i.e. only affecting
a certain API, as defined by the script logic options. This was already
possible with fbfrog in earlier versions, see e.g. 0ad30a46.

It's somewhat of a complication, but useful for some bindings, and also a
good chance to restructure the code a bit.