Fix C string literal lexing/parsing to allow <...> from macro expansion
This reduces the lexing of string literals to just reading in the whole
thing, and delays the parsing until later, to be done on-demand via the
new hStringLiteral(), which then takes care of evaluating escape sequences
etc.

This is very similar to the handling of number literals.

This is needed to support parsing #include "filenames" without evaluating
escape sequences.

This means that TK_STRING and the others now contain the whole string
literal, and as a result all code using it must be adjusted accordingly,
specifically: tkSpell() and CPP's stringification, and all string literal
parsing in the CPP/C parsers.

This has a nice side-effect: extra double quotes in CLI error messages are
gone.