console output: Only show a single, merged list of #included files
This changes the console output from showing the list of #included files
per API, as the files are being loaded, to only showing a single list at
the end, after all APIs were processed.

The single list should be nicer, because it's the merged version of all
the individual ones. API-specifics are easier to see, there should be much
less duplicate/redundant output.

With support for more targets there are more and more APIs, hence there was
more console output too, so reducing it becomes more important than before.