Improve #if generation: Don't produce directly nested VERORs/VERANDs
VERORs shouldn't be added as children of a VEROR, and VERANDs shouldn't
be added as children of a VERAND. astNewVEROR/astNewVERAND avoid that,
and this commit fixes a case where the #if condition simplification didn't
follow the rule.

Directly nested VERORs/VERANDs are useless, and can prevent other the
optimizations of hSimplify() from working, because they only check the
direct children of a VEROR, not considering/expecting nested VERORs.