Disallow multiple -declaredefines/-declareversions options
fbfrog does currently not handle conflicting -declareversions properly,
for example:

a.fbfrog:
	-declarebool A
	-ifdef A
		-declareversions V 1 2 3
	-else
		-declareversions V 1 2
	-endif

a.h:
	#if V == 1
		void v1(void);
	#else
		void v2plus(void);
	#endif

	#if V < 3
		void vlt3(void);
	#endif

a.bi:
	#if V = 1
		declare sub v1()
	#else
		declare sub v2plus()
	#endif

	declare sub vlt3()

That's wrong - the V<3 check is dropped completely because we have a
"V=1 or V=2" in frog.completeverors. But that only applies to some APIs,
not all.

It seems rather complex to fix this "properly", so I think for now it's
best to just disallow using multiple -declaredefines/-declareversions
options completely. It hasn't been needed for making fbbindings so far
anyways...