Change how -declareversions options are stored internally
The -declareversions option (if given) is now stored in a global array of
version numbers, and ASTNODEs representing checks against certain version
numbers etc. just store indices pointing into that array.

This makes the information about version numbers available to the #if
generation code, which can then do more optimizations in the future, for
example turning
	V=1 or V=2 or V=3
into
	V<=3
for -declareversions V 1 2 3 4.

Also, this allows some additional error checking in the CLI.