Show shorter target ids in console output, build target conditions later
Since each API now holds its os/arch, and because there now is -iftarget
that can be used instead of -ifdef to check for targets, the #if condition
generation code can build up the target checks as needed - it no longer
needs to be done at script evaluation time.

This also allows the console output to be simplified - the target can be
shown in a much more readable format.