Don't expand char/wchar_t typedefs affected by -nostring
Normally, fbfrog will expand char/wchar_t typedefs, so it can check whether
it's used as a pointer or array (then it should most likely be a string),
or as an individual char (then it should be an integer type, not a string
type).

-nostring prevents a declaration from being turned from integer type to
string type. This can also be applied to typedefs, which is useful for e.g.
winapi's CHAR/WCHAR typedefs that should be translated as byte/wchar_t.
And then -nostring should also prevent the expansion of these typedefs so
they'll actually be used. Otherwise we'd have to specify -nostring for all
the symbols where the typedefs were expanded.