CharStringPass pattern matching: Split into separate pass and set attribs
The pattern matching requires lots of context information, it's easier to
do that in a separate pass that marks matched nodes by setting some attrib
flags. Especially because we do isAffectedByNoString() on a looked-up
typedef, whose parent node is unknown at that point. Otherwise that typedef
may not be matched in some cases. (although typedefs usually aren't nested
anyways)