ast: Store origid in addition to id/alias to improve renaming
Previously, renaming was done by turning <id="a" alias=NULL> into
<id="b" alias="a">. I.e. the ASTNODE.alias field was used for two things:
holding the external name and the original C identifier.

This failed to handle cases where those two are different, which can
happen at least for declarations that have asm() aliases. And fbfrog also
sometimes wants to set aliases internally, without having them treated as
renames.

To fix this, ASTNODEs have to store an additional string -- the original C
identifier. This can be set separately from the alias if needed, and we
can also remove the NORENAMELIST flag and check for the origid to determine
whether a symbol was renamed.