Improve macro param conflict detection, add auto-renaming
1. Perfect macro param conflict detection: Will now check macro param
names against the actually emitted identifiers and FB keywords in each
macro body, as opposed to checking against all FB keywords etc. This
avoids plenty of false-positives. For this, the FB code emitting had to
be reimplemented to use a token buffer, which allows analyzing the
identifiers/keywords emitted for a specific macro's body.

2. Instead of adding TODO markers, affected macro params are now
auto-renamed, which is nicer than renaming manually, and in this specific
case, the auto-renaming is easy because the macro params are purely
internal/private symbols, not part of the binding's API.