Fix crash regression
AstNode.clone() could be called on a NULL THIS pointer and then crashed.

This is a regression from commit f0ec0ded, which replaced astClone(n)
and astCloneNode(n), which had the NULL checks, by clone() and
cloneNode() methods in the AstNode "class", which don't have the NULL
check anymore.