makefile: Add ENABLE_STANDALONE
Intent is to provide alternate build of fbfrog for some cooperation
with standalone installs of fbc.  This may allow for translation of
some C headers where a full toolchain and header installation is
either unnecessary for the translation, or there the end user is
managing the toolchains and paths themselves.

- make ENABLE_STANDALONE=1
  Enable search path for include files as follows:
  1. Absolute Path
  2. Current Directory
  3. <exepath>/include/fbfrog/
  4. <exepath>/fbfrog/include/

- make install ENABLE_STANDALONE=1 prefix=d:/<fbcdirectory>
  Copy 'fbfrog[.exe]' to prefix directory
  Copy 'include/fbfrog' to 'prefix/fbfrog/include'