Consider output path with bi destination file options
Change '-addinclude', '-emit', '-inclib', '-title', '-undef' options
to join the output path specified by '-o path' option to the .bi
destination filename.

If the -o specifies an output directory (not a file) and an option
specifies a destination bi file that is relative (not absolute),
then join the output path to the filename so that a later lookup
of the bi will succeed.

Without this change; if an output path is specified using -o option
and another option specifies a relative bi destination filename
then the lookup later fails unless the option also specifies the
output path.  Additionally, if the output path is set according to
a script, then the options specifying destination .bi files can't
be included in a fbfrog script without some preprocessing to set the
relative output path.