Strip single quoted arguments on win/dos
Single quoted arguments as might be done to
prevent wildcard expansion, prevent pattern
matching wildcards with filenames where some
arguments have single quotes and some don't.

for example:

... -include a.h -emit '*/a.h' a.bi
... -include 'a.h' -emit '*/a.h' a.bi

never matches a.h because on win the single
quotes are considered part of the filename.

When using fbfrog from command line, arguments
wrapped in single quotes are passed as-is and
the chances of having a filename wrapped in
single quotes seems rather slim.  But could
use "\'filename\'" if really wanted to force
the use.